feat: use reusable `Header` in `PostDetailsStack`
